+++ hooks/src/components/ImageGallery.tsx
@@ -1,34 +1,10 @@
-import React, { useState, useEffect } from 'react';
-
-const API_KEY = import.meta.env.VITE_PIXABAY_API_KEY;
-
-interface Image {
-  id: number;
-  webformatURL: string;
-  tags: string;
-}
+import React, { useState } from 'react';
+import useImageSearch from '../hooks/useImageSearch';
 
 const ImageGallery: React.FC = () => {
-  const [images, setImages] = useState<Image[]>([]);
   const [keyword, setKeyword] = useState('');
 
-  // useEffect를 사용하여 keyword가 변경될 때마다 이미지를 가져오는 로직을 작성해보세요.
-  useEffect(() => {
-    if (keyword) {
-      fetchImages().then((data) => {
-        setImages(data.hits);
-      });
-    }
-  }, [keyword]);
-
-  const fetchImages = async () => {
-    const requestURL =
-      'https://pixabay.com/api/?key=' + API_KEY + '&q=' + encodeURIComponent(keyword);
-
-    const response = await fetch(requestURL);
-
-    return response.json();
-  };
+  const { images } = useImageSearch(keyword);
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
